Added login-save function

+++ app/src/main/java/calculinc/google/httpssites/skol_app/MainActivity.java
@@ -1,10 +1,18 @@
 package calculinc.google.httpssites.skol_app;
 
+import android.content.ClipData;
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.Paint;
+import android.graphics.PorterDuff;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
+import android.support.v7.view.menu.MenuView;
+import android.text.TextPaint;
 import android.util.Log;
 import android.view.View;
 import android.support.design.widget.NavigationView;
@@ -19,18 +27,25 @@ import android.view.ViewDebug;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.ViewFlipper;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
+
+import static calculinc.google.httpssites.skol_app.R.drawable.rainbowcolor;
 
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
@@ -38,7 +53,10 @@ public class MainActivity extends AppCompatActivity
     long id_number;
     int day = GregorianCalendar.getInstance().get(Calendar.DAY_OF_WEEK);
     int week = GregorianCalendar.getInstance().get(Calendar.WEEK_OF_YEAR);
+    String pdfFileName = "Nova.pdf";
     File pdfFile;
+    String loginFileName = "login.txt";
+    File loginFile;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,15 +86,56 @@ public class MainActivity extends AppCompatActivity
         TextView kek = (TextView) findViewById(R.id.öpö);
         kek.setText(String.valueOf(week));
 
-        pdfFile.delete();
+        pdfFile = new File(getFilesDir() + "/" + pdfFileName);
+
+        if (pdfFile.exists()) {
+            pdfFile.delete();
+        }
+
+        loginFile = new File(getFilesDir() + "/" + loginFileName);
+
+        try {
+            if (loginFile.exists()) {
+                FileInputStream fis = openFileInput(loginFileName);
+                InputStreamReader isr = new InputStreamReader(fis);
+                BufferedReader bufferedReader = new BufferedReader(isr);
+                StringBuilder sb = new StringBuilder();
+                String line;
+                while ((line = bufferedReader.readLine()) != null) {
+                    sb.append(line);
+                }
+                String loginString = sb.toString();
+
+                id_number = Long.parseLong(loginString);
+                EditText personal_id = (EditText) findViewById(R.id.editText);
+                if (id_number == 0) {
+                    personal_id.setText("");
+                } else {
+                    personal_id.setText(String.valueOf(id_number));
+                }
+
+            }
+        }catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     public void loginSubmit (View view) {
         final EditText personal_id = (EditText) findViewById(R.id.editText);
+
         if (personal_id.getText().toString().equals("")) {
             id_number = 0;
         } else {
             id_number = Long.parseLong(personal_id.getText().toString());
+        }
+        String login_content = String.valueOf(id_number);
+        FileOutputStream outputStream;
+        try {
+            outputStream = openFileOutput(loginFileName,MODE_PRIVATE);
+            outputStream.write(login_content.getBytes());
+            outputStream.close();
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 
@@ -90,12 +149,16 @@ public class MainActivity extends AppCompatActivity
         }
     }
 
+    private Menu menu;
+
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.main, menu);
+
         return true;
     }
+
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -129,12 +192,13 @@ public class MainActivity extends AppCompatActivity
 
             toolbar.setTitle(R.string.Tab_2);
             vf.setDisplayedChild(1);
-            new DownloadFile().execute("http://www.novasoftware.se/ImgGen/schedulegenerator.aspx?format=pdf&schoolid=81530/sv-se&type=0&id=" + id_number  + "&period=&week=" + 20 + "&mode=0&printer=0&colors=32&head=5&clock=7&foot=1&day=" + 0 + "&width=400&height=640","Nova.pdf");
+            new DownloadFile().execute("http://www.novasoftware.se/ImgGen/schedulegenerator.aspx?format=pdf&schoolid=81530/sv-se&type=0&id=" + id_number  + "&period=&week=" + 20 + "&mode=0&printer=0&colors=32&head=5&clock=7&foot=1&day=" + 0 + "&width=400&height=640");
 
         } else if (id == R.id.nav_matsedel) {
 
             toolbar.setTitle(R.string.Tab_3);
             vf.setDisplayedChild(2);
+            ggfunktion();
 
         } else if (id == R.id.nav_laxor) {
 
@@ -157,16 +221,64 @@ public class MainActivity extends AppCompatActivity
         return true;
     }
 
+    public void drawing()  {
+
+        Paint paint = new Paint();
+        paint.setAntiAlias(true);
+        Paint plebpaint = new Paint();
+        plebpaint.setColor(Color.parseColor("#ffffff"));
+        plebpaint.setStyle(Paint.Style.STROKE);
+        paint.setColor(Color.parseColor("#61b0ae"));
+        TextPaint textpaint = new TextPaint();
+        textpaint.setColor(Color.parseColor("#ffffff"));
+        textpaint.setAntiAlias(true);
+
+        Bitmap bg = Bitmap.createBitmap(480,800, Bitmap.Config.ARGB_8888);
+        Canvas canvas = new Canvas(bg);
+        canvas.drawRect(12,120,468,250,paint);
+        canvas.drawRect(17,125,463,245,plebpaint);
+        textpaint.setTextSize(32);
+        canvas.drawText("( ͡° ͜ʖ ͡°)",120,200,textpaint);
+        ImageView föfan = (ImageView) findViewById(R.id.föfan);
+        föfan.setImageBitmap(bg);
+    }
+
+    public void ggfunktion() {
+
+        Button button = (Button) findViewById(R.id.plebknapp);
+        final LinearLayout linearlayout = (LinearLayout) findViewById(R.id.linearlayout);
+        final TextView ggboistext = (TextView) findViewById(R.id.gg_bois);
+        final ImageView hexa = (ImageView) findViewById(R.id.hexagon);
+        final TextView clock = (TextView) findViewById(R.id.klockan);
+        final LinearLayout midbox = (LinearLayout) findViewById(R.id.mid_box);
+        final ImageView hexa2 = (ImageView) findViewById(R.id.hexagon2);
+
+        button.setOnClickListener(new View.OnClickListener() {
+            int status = 1;
+            int test = 1;
+            @Override
+            public void onClick(View view) {
+
+                ggboistext.setText("GG BOIS");
+                linearlayout.setBackgroundResource(rainbowcolor);
+                hexa.setColorFilter(Color.parseColor("#16aba0"), PorterDuff.Mode.SRC_ATOP);
+                hexa2.setColorFilter(Color.parseColor("#16aba0"), PorterDuff.Mode.SRC_ATOP);
+                clock.setText(String.valueOf(GregorianCalendar.getInstance().get(Calendar.HOUR_OF_DAY)) + ":" + String.valueOf(GregorianCalendar.getInstance().get(Calendar.MINUTE)));
+                midbox.setBackgroundColor(Color.parseColor("#16aba0"));
+                midbox.getLayoutParams().height = ((int) getResources().getDimension(R.dimen.dp_unit))*180;
+                midbox.requestLayout();
+            }
+        });
+    }
+
     private class DownloadFile extends AsyncTask<String, Void, Void>{
 
         @Override
         protected Void doInBackground(String... strings) {
             String fileUrl = strings[0];
-            String fileName = strings[1];
+            //String fileName = strings[1];
 
             final int  MEGABYTE = 1024 * 1024;
-
-            pdfFile = new File(getFilesDir() + "/" + fileName);
 
             String lel = "finns";
             try {
@@ -181,7 +293,7 @@ public class MainActivity extends AppCompatActivity
                     pdfFile.delete();
                     lel = "borta";
                 }
-                FileOutputStream fileOutputStream = openFileOutput(fileName,MODE_PRIVATE);
+                FileOutputStream fileOutputStream = openFileOutput(pdfFileName,MODE_PRIVATE);
                 int totalSize = urlConnection.getContentLength();
                 lel = "finns";
 
@@ -221,6 +333,7 @@ public class MainActivity extends AppCompatActivity
             } else {
                 kek = "pdfFile is short " + pdfFile.length();
             }
+            //delete when not needed
 
             return null;
         }
